Rename @atlaspack/eslint-plugin-public to @atlaspack/eslint-plugin
