+++ packages/core/eslint-plugin/CHANGELOG.md
@@ -1,0 +1,1 @@
+# @atlaspack/eslint-plugin
