+++ packages/core/eslint-plugin/src/rules/importcond-type-annotations/index.ts
