+++ packages/core/eslint-plugin/src/utils/index.ts
@@ -7,5 +7,5 @@ export interface PluginDocs {
 
 export const createRule = ESLintUtils.RuleCreator<PluginDocs>(
   (name) =>
-    `https://github.com/atlassian-labs/atlaspack/tree/main/packages/core/eslint-plugin-public/src/rules/${name}/README.md`,
+    `https://github.com/atlassian-labs/atlaspack/tree/main/packages/core/eslint-plugin/src/rules/${name}/README.md`,
 );
