+++ packages/core/eslint-plugin/tsconfig.json
