+++ packages/dev/eslint-config/CHANGELOG.md
@@ -20,7 +20,7 @@
 - [#414](https://github.com/atlassian-labs/atlaspack/pull/414) [`eff9809`](https://github.com/atlassian-labs/atlaspack/commit/eff98093703b9999a511b87a19562f5aaccfcb53) Thanks [@alshdavid](https://github.com/alshdavid)! - Added type:commonjs to package.json files
 
 - Updated dependencies [[`bfe81e5`](https://github.com/atlassian-labs/atlaspack/commit/bfe81e551c4e4bb2cac7fc4745222e66962c1728), [`eff9809`](https://github.com/atlassian-labs/atlaspack/commit/eff98093703b9999a511b87a19562f5aaccfcb53)]:
-  - @atlaspack/eslint-plugin@2.14.0
+  - @atlaspack/eslint-plugin-internal@2.14.0
 
 ## 2.13.1
 
@@ -29,7 +29,7 @@
 - [`3ddd868`](https://github.com/atlassian-labs/atlaspack/commit/3ddd8682a6edb5c6a35357cfa3ade5741aff5f06) Thanks [@yamadapc](https://github.com/yamadapc)! - Add identifier registry and VCS tracing
 
 - Updated dependencies [[`3ddd868`](https://github.com/atlassian-labs/atlaspack/commit/3ddd8682a6edb5c6a35357cfa3ade5741aff5f06)]:
-  - @atlaspack/eslint-plugin@2.13.1
+  - @atlaspack/eslint-plugin-internal@2.13.1
 
 ## 2.13.0
 
@@ -40,4 +40,4 @@
 ### Patch Changes
 
 - Updated dependencies [[`b4dbd4d`](https://github.com/atlassian-labs/atlaspack/commit/b4dbd4d5b23d1b7aa3fcdf59cc7bc8bedd3a59cf)]:
-  - @atlaspack/eslint-plugin@2.13.0
+  - @atlaspack/eslint-plugin-internal@2.13.0
