+++ packages/dev/eslint-config/package.json
@@ -6,7 +6,7 @@
     "@babel/eslint-parser": "^7.22.11",
     "@babel/plugin-syntax-flow": "^7.14.5",
     "@babel/plugin-transform-react-jsx": "^7.14.9",
-    "@atlaspack/eslint-plugin": "2.14.1",
+    "@atlaspack/eslint-plugin-internal": "2.14.1",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-flowtype": "^8.0.3",
     "eslint-plugin-import": "^2.22.1",
