+++ packages/dev/eslint-plugin-internal/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@atlaspack/eslint-plugin",
+  "name": "@atlaspack/eslint-plugin-internal",
   "private": true,
   "version": "2.14.1",
   "main": "index.js",
