+++ packages/dev/eslint-plugin-internal/src/rules/no-ff-module-level-eval.js
