+++ packages/dev/eslint-plugin-internal/src/rules/no-relative-import.js
