+++ packages/dev/eslint-plugin-internal/src/rules/no-self-package-imports.js
