+++ packages/dev/eslint-plugin-internal/src/utils.js
